requirements.txt
# Data Management Service - Python Dependencies
# Enterprise Django REST Framework Application

# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.9

# AWS S3 Storage
boto3>=1.34.0
django-storages>=1.14.0

# JWT Authentication
PyJWT>=2.8.0
cryptography>=41.0.0

# CORS Support
django-cors-headers>=4.3.0

# Environment Variables
python-decouple>=3.8

# Data Validation
jsonschema>=4.20.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development Tools (optional)
django-extensions>=3.2.3
